requirements.txt
affine==2.2.2
asynchronousfilereader==0.2.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
configparser==3.7.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
GDAL==2.3.3
geographiclib==1.49
geojson==2.4.1
geomet==0.2.0.post2
geopy==1.20.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-storage==1.17.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
html2text==2018.1.9
idna==2.8
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.16.4
olefile==0.46
packaging==19.0
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
Pygments==2.4.2
pyparsing==2.4.0
pyproj==2.2.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-logstash==0.4.6
pytz==2019.2
pyzmq==18.0.0
qtconsole==4.5.2
rasterio==1.0.21
requests==2.22.0
rsa==4.0
sedas-pyapi==0.4.1
Send2Trash==1.5.0
sentinelsat==0.13
Shapely==1.6.4.post2
six==1.12.0
snuggs==1.4.6
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1